Split unused file, tag and cluster entity deletion into a SELECT ... FOR UPDATE and a DELETE to prevent database locking

+++ genie-web/src/main/java/com/netflix/genie/web/services/FileService.java
@@ -47,5 +47,5 @@ public interface FileService {
      *                         will be deleted. Inclusive
      * @return The number of files deleted
      */
-    int deleteUnusedFiles(@NotNull final Instant createdThreshold);
+    long deleteUnusedFiles(@NotNull final Instant createdThreshold);
 }
