+++ genie-web/src/main/java/com/netflix/genie/web/tasks/leader/DatabaseCleanupTask.java
@@ -154,7 +154,7 @@ public class DatabaseCleanupTask extends LeadershipTask {
                 log.debug("Skipping clusters cleanup");
                 this.numDeletedClusters.set(0);
             } else {
-                final int countDeletedClusters = this.clusterService.deleteTerminatedClusters();
+                final long countDeletedClusters = this.clusterService.deleteTerminatedClusters();
                 log.info(
                     "Deleted {} clusters that were in TERMINATED state and weren't attached to any jobs",
                     countDeletedClusters
@@ -169,7 +169,7 @@ public class DatabaseCleanupTask extends LeadershipTask {
                 log.debug("Skipping files cleanup");
                 this.numDeletedFiles.set(0);
             } else {
-                final int countDeletedFiles = this.fileService.deleteUnusedFiles(creationThreshold);
+                final long countDeletedFiles = this.fileService.deleteUnusedFiles(creationThreshold);
                 log.info(
                     "Deleted {} files that were unused by any resource and created over an hour ago",
                     countDeletedFiles
@@ -181,7 +181,7 @@ public class DatabaseCleanupTask extends LeadershipTask {
                 log.debug("Skipping tags cleanup");
                 this.numDeletedTags.set(0);
             } else {
-                final int countDeletedTags = this.tagService.deleteUnusedTags(creationThreshold);
+                final long countDeletedTags = this.tagService.deleteUnusedTags(creationThreshold);
                 log.info(
                     "Deleted {} tags that were unused by any resource and created over an hour ago",
                     countDeletedTags
